Simplified the arguments to many OOFCanvas functions for simplicity.
Many functions had separate x and y arguments instead of just accepting
an OOFCanvas::Coord.  This was done thinking that it would make it
simpler to use OOFCanvas in an environment that already had its own
Coord class (like OOF2).  But OOFCanvas::Coord was already used in
the OOFCanvas API in other places, so it was silly to avoid it.
Now the OOFCanvas Coord, ICoord, and Color classes are used as
arguments where it makes sense.

Added a canvasColor() function in ccolor.C that converts an OOF2
CColor to an OOFCanvas::Color.  There already was an analogous
function in Python.

When returned from C++ to Python, OOFCanvas::Coords and ICoords
are converted to simple tuples by a typemap.  This allows their
x and y components to be easily accessed with  __getitem__, or
allows them to be easily used in constructors for some other type,
  pt = Point(*canvasMethodReturningACoord())

In the other direction, coordinate-like objects passed in to
C++ from Python are converted into Coords or ICoords (depending
on the typemap) as long as they support __getitem__, so that
obj[0] is the x component and obj[1] is the y component.

+++ SRC/common/IO/OOFCANVAS/canvas.h
@@ -85,10 +85,8 @@ namespace OOFCanvas {
     double user2pixel(double) const;
     double pixel2user(double) const;
 
-    // This version just exists for calling from Python. Used when
-    // converting GdkEvent coordinates to user coordinates.  Event
-    // coords are floats.
-    Coord *pixel2user(double, double) const;
+    // This version just exists for calling from Python.
+    Coord *pixel2user(int, int) const;
 
     void setAntialias(bool);
     void setMargin(double);
